# Defer allocations and optimize BPE encoding.