--- a/tokenizers/src/models/bpe/word.rs
+++ b/tokenizers/src/models/bpe/word.rs
@@ -78,6 +78,12 @@ impl Word {
         Word { symbols: vec![] }
     }
 
+    pub(super) fn with_capacity(capacity: usize) -> Self {
+        Word {
+            symbols: Vec::with_capacity(capacity),
+        }
+    }
+
     pub(super) fn add(&mut self, c: u32) {
         let (prev, next) = {
             let len = self.symbols.len() as isize;
@@ -142,118 +148,114 @@ impl Word {
     }
 
     pub(super) fn merge_all(&mut self, merges: &HashMap<Pair, (u32, u32)>, dropout: Option<f32>) {
-        let dropout = dropout.unwrap_or(0.0);
         let mut queue = BinaryHeap::with_capacity(self.symbols.len());
         let mut skip = Vec::with_capacity(queue.len());
 
-        self.symbols
-            .windows(2)
-            .enumerate()
-            .for_each(|(index, window)| {
-                let pair = (window[0].c, window[1].c);
-                if let Some(m) = merges.get(&pair) {
-                    queue.push(Merge {
+        queue.extend(
+            self.symbols
+                .windows(2)
+                .enumerate()
+                .filter_map(|(index, window)| {
+                    let pair = (window[0].c, window[1].c);
+                    merges.get(&pair).map(|m| Merge {
                         pos: index,
                         rank: m.0,
                         new_id: m.1,
-                    });
-                }
-            });
+                    })
+                }),
+        );
 
         while let Some(top) = queue.pop() {
-            if dropout > 0.0 && thread_rng().gen::<f32>() < dropout {
+            if dropout
+                .map(|d| thread_rng().gen::<f32>() < d)
+                .unwrap_or(false)
+            {
                 skip.push(top);
             } else {
                 // Re-insert the skipped elements
-                skip.drain(..).for_each(|s| queue.push(s));
+                queue.extend(skip.drain(..));
 
-                if self.symbols[top.pos].len > 0 {
-                    // Do nothing if we are the last symbol
-                    if self.symbols[top.pos].next == -1 {
-                        continue;
+                if self.symbols[top.pos].len == 0 {
+                    continue;
+                }
+                // Do nothing if we are the last symbol
+                if self.symbols[top.pos].next == -1 {
+                    continue;
+                }
+
+                let next_pos = self.symbols[top.pos].next as usize;
+                let right = self.symbols[next_pos];
+
+                // Make sure we are not processing an expired queue entry
+                let target_new_pair = (self.symbols[top.pos].c, right.c);
+                if !merges
+                    .get(&target_new_pair)
+                    .map_or(false, |(_, new_id)| *new_id == top.new_id)
+                {
+                    continue;
+                }
+
+                // Otherwise, let's merge
+                self.symbols[top.pos].merge_with(&right, top.new_id);
+                // Tag the right part as removed
+                self.symbols[next_pos].len = 0;
+
+                // Update `prev` on the new `next` to the current pos
+                if right.next > -1 && (right.next as usize) < self.symbols.len() {
+                    self.symbols[right.next as usize].prev = top.pos as isize;
+                }
+
+                // Insert the new pair formed with the previous symbol
+                let current = &self.symbols[top.pos];
+                if current.prev >= 0 {
+                    let prev = current.prev as usize;
+                    let prev_symbol = self.symbols[prev];
+                    let new_pair = (prev_symbol.c, current.c);
+                    if let Some((rank, new_id)) = merges.get(&new_pair) {
+                        queue.push(Merge {
+                            pos: current.prev as usize,
+                            rank: *rank,
+                            new_id: *new_id,
+                        });
                     }
+                }
 
-                    let next_pos = self.symbols[top.pos].next as usize;
-                    let right = self.symbols[next_pos];
-
-                    // Make sure we are not processing an expired queue entry
-                    let target_new_pair = (self.symbols[top.pos].c, right.c);
-                    if !merges
-                        .get(&target_new_pair)
-                        .map_or(false, |(_, new_id)| *new_id == top.new_id)
-                    {
-                        continue;
-                    }
-
-                    // Otherwise, let's merge
-                    self.symbols[top.pos].merge_with(&right, top.new_id);
-                    // Tag the right part as removed
-                    self.symbols[next_pos].len = 0;
-
-                    // Update `prev` on the new `next` to the current pos
-                    if right.next > -1 && (right.next as usize) < self.symbols.len() {
-                        self.symbols[right.next as usize].prev = top.pos as isize;
-                    }
-
-                    // Insert the new pair formed with the previous symbol
-                    let current = &self.symbols[top.pos];
-                    if current.prev >= 0 {
-                        let prev = current.prev as usize;
-                        let prev_symbol = self.symbols[prev];
-                        let new_pair = (prev_symbol.c, current.c);
-                        if let Some((rank, new_id)) = merges.get(&new_pair) {
-                            queue.push(Merge {
-                                pos: current.prev as usize,
-                                rank: *rank,
-                                new_id: *new_id,
-                            });
-                        }
-                    }
-
-                    // Insert the new pair formed with the next symbol
-                    let next = current.next as usize;
-                    if next < self.symbols.len() {
-                        let next_symbol = self.symbols[next];
-                        let new_pair = (current.c, next_symbol.c);
-                        if let Some((rank, new_id)) = merges.get(&new_pair) {
-                            queue.push(Merge {
-                                pos: top.pos,
-                                rank: *rank,
-                                new_id: *new_id,
-                            });
-                        }
+                // Insert the new pair formed with the next symbol
+                let next = current.next as usize;
+                if next < self.symbols.len() {
+                    let next_symbol = self.symbols[next];
+                    let new_pair = (current.c, next_symbol.c);
+                    if let Some((rank, new_id)) = merges.get(&new_pair) {
+                        queue.push(Merge {
+                            pos: top.pos,
+                            rank: *rank,
+                            new_id: *new_id,
+                        });
                     }
                 }
             }
         }
 
         // Filter out the removed symbols
-        let mut i = 0;
-        loop {
-            if self.symbols.len() <= i {
-                break;
-            }
-
-            if self.symbols[i].len == 0 {
-                self.symbols.remove(i);
-            } else {
-                i += 1;
-            }
-        }
+        self.symbols.retain(|s| s.len != 0);
     }
 
     pub(super) fn get_chars(&self) -> Vec<u32> {
         self.symbols.iter().map(|s| s.c).collect()
     }
 
-    pub(super) fn get_offsets(&self) -> Vec<(usize, usize)> {
-        let mut offsets = vec![];
+    pub(super) fn get_chars_iter<'a>(&'a self) -> impl Iterator<Item = u32> + 'a {
+        self.symbols.iter().map(|s| s.c)
+    }
+
+    pub(super) fn get_offsets_iter<'a>(&'a self) -> impl Iterator<Item = (usize, usize)> + 'a {
         let mut pos = 0;
-        for symbol in &self.symbols {
-            offsets.push((pos, pos + symbol.len));
-            pos += symbol.len;
-        }
-        offsets
+        self.symbols.iter().map(move |symbol| {
+            let new_pos = pos + symbol.len;
+            let offset = (pos, new_pos);
+            pos = new_pos;
+            offset
+        })
     }
 }