Improvements over `master` on my idling 2700x:

~~~bash
$ make bench
[...]
BPE GPT2 encode         time:   [35.912 us 37.428 us 38.737 us]
                        change: [-17.820% -12.223% -6.7713%] (p = 0.00 < 0.05)
                        Performance has improved.
slope  [35.912 us 38.737 us] R^2            [0.8299894 0.8384948]
mean   [34.624 us 38.082 us] std. dev.      [2.9044 us 4.8232 us]
median [34.415 us 38.963 us] med. abs. dev. [1.4156 us 6.9376 us]
[...]
BPE GPT2 encode batch   time:   [4.5797 ms 4.6026 ms 4.6260 ms]
                        change: [-5.8516% -5.0030% -4.2109%] (p = 0.00 < 0.05)
                        Performance has improved.
slope  [4.5797 ms 4.6260 ms] R^2            [0.9959612 0.9959216]
mean   [4.5750 ms 4.6275 ms] std. dev.      [44.414 us 72.872 us]
median [4.5675 ms 4.6303 ms] med. abs. dev. [24.338 us 105.32 us]
[...]
BPE GPT2 encode, no cache
                        time:   [47.176 us 47.681 us 48.237 us]
                        change: [-17.601% -16.125% -14.420%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 20 measurements (5.00%)
  1 (5.00%) high severe
slope  [47.176 us 48.237 us] R^2            [0.9816144 0.9809403]
mean   [46.974 us 48.175 us] std. dev.      [610.63 ns 2.0715 us]
median [46.703 us 47.610 us] med. abs. dev. [353.26 ns 1.3329 us]
[...]
BPE GPT2 encode batch, no cache
                        time:   [5.6339 ms 5.6701 ms 5.7115 ms]
                        change: [-14.359% -13.290% -12.345%] (p = 0.00 < 0.05)
                        Performance has improved.
slope  [5.6339 ms 5.7115 ms] R^2            [0.9932644 0.9928743]
mean   [5.6197 ms 5.7071 ms] std. dev.      [70.391 us 125.93 us]
median [5.6141 ms 5.7190 ms] med. abs. dev. [44.187 us 171.61 us]
[...]
BPE Train vocabulary (small)
                        time:   [17.015 ms 17.102 ms 17.225 ms]
                        change: [-0.0733% +0.7268% +1.5477%] (p = 0.11 > 0.05)
                        No change in performance detected.
slope  [17.015 ms 17.225 ms] R^2            [0.9981535 0.9976542]
mean   [16.990 ms 17.205 ms] std. dev.      [93.667 us 231.31 us]
median [16.954 ms 17.231 ms] med. abs. dev. [32.239 us 305.75 us]
[...]
BPE Train vocabulary (big)
                        time:   [708.27 ms 721.25 ms 734.64 ms]
                        change: [-2.2477% +0.3769% +3.1911%] (p = 0.79 > 0.05)
                        No change in performance detected.
mean   [708.27 ms 734.64 ms] std. dev.      [13.262 ms 27.204 ms]
median [702.05 ms 743.33 ms] med. abs. dev. [5.6081 ms 39.705 ms]
~~~